+++ src/refactorings/flip-ternary/flip-ternary.ts
@@ -2,7 +2,7 @@ import { Editor, ErrorReason } from "../../editor/editor";
 import { Selection } from "../../editor/selection";
 import * as t from "../../ast";
 
-import { getNegatedBinaryOperator } from "../negate-expression/negate-expression";
+import { getNegatedBinaryOperator } from "../invert-boolean-logic/invert-boolean-logic";
 
 export { flipTernary, createVisitor as hasTernaryToFlip };