+++ src/refactorings/invert-boolean-logic/index.ts
@@ -1,25 +1,25 @@
 import {
   canNegateExpression,
-  negateExpression,
+  invertBooleanLogic,
   getNegatedOperator
-} from "./negate-expression";
+} from "./invert-boolean-logic";
 
 import { RefactoringWithActionProvider } from "../../types";
 
 const config: RefactoringWithActionProvider = {
   command: {
-    key: "negateExpression",
-    operation: negateExpression,
-    title: "Negate Expression"
+    key: "invertBooleanLogic",
+    operation: invertBooleanLogic,
+    title: "Invert Boolean Logic (De Morgan's Law)"
   },
   actionProvider: {
-    message: "Negate the expression",
+    message: "Invert boolean logic (De Morgan's law)",
     createVisitor: canNegateExpression,
     updateMessage(path) {
       const operator = getNegatedOperator(path.node);
       return operator
-        ? `Negate the expression (use ${operator} instead)`
-        : "Negate the expression";
+        ? `Invert boolean logic (use ${operator} instead) (De Morgan's law)`
+        : "Invert boolean logic (De Morgan's law)";
     }
   }
 };